// frontend/src/app/types/ranking.ts
export type RankingGeralItem = {
  nome: string;
  pontos: number;
  acertos_placar: number;
  acertos_saldo: number;
  acertos_resultado: number;
  erros: number;
  aproveitamento: number;
  perc_placar: number;
  perc_saldo: number;
  perc_resultado: number;
};

export type RankingRodadaItem = {
  nome: string;
  pontos: number;
  acertos_placar: number;
  acertos_saldo: number;
  acertos_resultado: number;
};